// packages/api/src/generated/admin/models/adminJobDetailResponseRecruitmentStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */

/**
 * 저장하지 않고 마감 처리 일시와 모집 종료 일시로 계산한 값입니다.
 *
 * | 값 | code | 설명 |
 * | --- | --- | --- |
 * | `RECRUITING` | 1 | 모집 중 |
 * | `CLOSED` | 2 | 모집 마감 |
 */
export type AdminJobDetailResponseRecruitmentStatus = typeof AdminJobDetailResponseRecruitmentStatus[keyof typeof AdminJobDetailResponseRecruitmentStatus];


export const AdminJobDetailResponseRecruitmentStatus = {
  RECRUITING: 'RECRUITING',
  CLOSED: 'CLOSED',
} as const;