// packages/api/src/generated/admin/models/adminJobDetailResponseReviewStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */

/**
 * 크롤링 수집분은 null입니다.
 *
 * | 값 | code | 설명 |
 * | --- | --- | --- |
 * | `PENDING` | 1 | 검수 대기 |
 * | `APPROVED` | 2 | 승인 |
 * | `REJECTED` | 3 | 반려 |
 */
export type AdminJobDetailResponseReviewStatus = typeof AdminJobDetailResponseReviewStatus[keyof typeof AdminJobDetailResponseReviewStatus];


export const AdminJobDetailResponseReviewStatus = {
  PENDING: 'PENDING',
  APPROVED: 'APPROVED',
  REJECTED: 'REJECTED',
} as const;