// packages/api/src/generated/admin/models/adminJobDetailResponseVisibility.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Ogonggo Admin API
 * OpenAPI spec version: v1
 */

/**
 * | 값 | code | 설명 |
 * | --- | --- | --- |
 * | `VISIBLE` | 1 | 노출 |
 * | `HIDDEN` | 2 | 비노출 |
 */
export type AdminJobDetailResponseVisibility = typeof AdminJobDetailResponseVisibility[keyof typeof AdminJobDetailResponseVisibility];


export const AdminJobDetailResponseVisibility = {
  VISIBLE: 'VISIBLE',
  HIDDEN: 'HIDDEN',
} as const;